remove modular plugin feature set to be exe installable update wordlist
