+++ pycep/data/word_list.txt
@@ -4225,3 +4225,18 @@ xff
 __1
 _1
 ___1
+raw_input
+wouldn
+string_1
+string_2
+string_3
+string_4
+string_5
+string_6
+string_7
+string_8
+string_9
+substring
+indexerror
+my_string
+hasn
